+++ daybed/tests/test_backend.py
@@ -108,7 +108,7 @@ class BackendTestBase(object):
         self._create_model()
         self.db.put_record('modelname', self.record, ['Alexis'])
 
-        # When we put a new version of a data item, we should keep the list of
+        # When we put a new version of a record, we should keep the list of
         # authors.
         item_id = self.db.put_record('modelname', self.record, ['Remy'])
         self.db.put_record('modelname', self.record, ['Alexis'], item_id)